fix: RefSectionHandler to support mgmt-api/api sections
